avm2: `graphics` is a `Sprite` property, not a `MovieClip` property
