+++ core/src/avm2/globals/flash/display/sprite.rs
@@ -1,10 +1,11 @@
 //! `flash.display.Sprite` builtin/prototype
 
 use crate::avm2::activation::Activation;
-use crate::avm2::class::Class;
+use crate::avm2::class::{Class, ClassAttributes};
 use crate::avm2::method::Method;
 use crate::avm2::names::{Namespace, QName};
-use crate::avm2::object::Object;
+use crate::avm2::object::{Object, StageObject, TObject};
+use crate::avm2::traits::Trait;
 use crate::avm2::value::Value;
 use crate::avm2::Error;
 use gc_arena::{GcCell, MutationContext};
@@ -31,9 +32,29 @@ pub fn class_init<'gc>(
     Ok(Value::Undefined)
 }
 
+/// Implements `graphics`.
+pub fn graphics<'gc>(
+    activation: &mut Activation<'_, 'gc, '_>,
+    this: Option<Object<'gc>>,
+    _args: &[Value<'gc>],
+) -> Result<Value<'gc>, Error> {
+    if let Some(dobj) = this.and_then(|o| o.as_display_object()) {
+        let graphics_proto = activation.context.avm2.prototypes().graphics;
+
+        return Ok(StageObject::for_display_object(
+            activation.context.gc_context,
+            dobj,
+            graphics_proto,
+        )
+        .into());
+    }
+
+    Ok(Value::Undefined)
+}
+
 /// Construct `Sprite`'s class.
 pub fn create_class<'gc>(mc: MutationContext<'gc, '_>) -> GcCell<'gc, Class<'gc>> {
-    Class::new(
+    let class = Class::new(
         QName::new(Namespace::package("flash.display"), "Sprite"),
         Some(
             QName::new(
@@ -45,5 +66,16 @@ pub fn create_class<'gc>(mc: MutationContext<'gc, '_>) -> GcCell<'gc, Class<'gc>
         Method::from_builtin(instance_init),
         Method::from_builtin(class_init),
         mc,
-    )
+    );
+
+    let mut write = class.write(mc);
+
+    write.set_attributes(ClassAttributes::SEALED);
+
+    write.define_instance_trait(Trait::from_getter(
+        QName::new(Namespace::public(), "graphics"),
+        Method::from_builtin(graphics),
+    ));
+
+    class
 }
